Start converting VisionSource/LightSource/etc to new V12 structure
