+++ src/module/canvas/perception/hearing-source.ts
@@ -1,8 +1,8 @@
 import { TokenPF2e } from "../token/index.ts";
 
 /** A `PointSource` to track token hearing sense */
-class HearingSource<TObject extends TokenPF2e> extends SoundSource<TObject> {
-    protected override _initialize(data: PointSourceData): void {
+class HearingSource<TObject extends TokenPF2e> extends foundry.canvas.sources.PointSoundSource<TObject> {
+    protected override _initialize(data: PointEffectSourceData): void {
         super._initialize(data);
         this.data.x = this.object.center.x;
         this.data.y = this.object.center.y;
