+++ types/foundry/client-esm/canvas/sources/module.ts
@@ -1,0 +1,9 @@
+export { default as BaseEffectSource } from "./base-effect-source.ts";
+export { default as BaseLightSource } from "./base-light-source.ts";
+export { default as GlobalLightSource } from "./global-light-source.ts";
+export { default as PointDarknessSource } from "./point-darkness-source.ts";
+export { default as PointLightSource } from "./point-light-source.ts";
+export { default as PointMovementSource } from "./point-movement-source.ts";
+export { default as PointSoundSource } from "./point-sound-source.ts";
+export { default as PointVisionSource } from "./point-vision-source.ts";
+export { default as RenderedEffectSource } from "./rendered-effect-source.ts";
