+++ types/foundry/client-esm/canvas/sources/point-darkness-source.d.ts
@@ -1,0 +1,5 @@
+import type { PointEffectLightSource } from "./point-effect-source-mixes.ts";
+
+export default class PointDarknessSource<
+    TObject extends Token | AmbientLight,
+> extends PointEffectLightSource<TObject> {}
