+++ types/foundry/client-esm/canvas/sources/point-light-source.d.ts
@@ -1,0 +1,3 @@
+import type { PointEffectLightSource } from "./point-effect-source-mixes.d.ts";
+
+export default class PointLightSource<TObject extends Token | AmbientLight> extends PointEffectLightSource<TObject> {}
