+++ types/foundry/client/pixi/module.d.ts
@@ -1,2 +1,3 @@
 export * as edges from "./edges/module.ts";
+export * as sources from "../../client-esm/canvas/sources/module.ts";
 export * from "./scene-manager.ts";
