--- types/foundry/client/pixi/sources/index.d.ts
@@ -1,5 +1,0 @@
-import "./base-source.d.ts";
-import "./light-source.d.ts";
-import "./rendered-source.d.ts";
-import "./sound-source.d.ts";
-import "./vision-source.d.ts";
